+++ src/main/java/com/samcomo/dbz/member/jwt/filter/JwtFilter.java
@@ -1,11 +1,11 @@
-package com.samcomo.dbz.member.model.jwt.filter;
+package com.samcomo.dbz.member.jwt.filter;
 
 import static com.samcomo.dbz.member.model.constants.TokenType.ACCESS_TOKEN;
 
 import com.samcomo.dbz.member.model.constants.MemberRole;
 import com.samcomo.dbz.member.model.dto.MemberDetails;
 import com.samcomo.dbz.member.model.entity.Member;
-import com.samcomo.dbz.member.model.jwt.JwtUtil;
+import com.samcomo.dbz.member.jwt.JwtUtil;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
